+++ src/config.rs
@@ -63,6 +63,8 @@ impl AppService {
         (self.config, self.services)
     }
 
+    /// Clones inner config and default service, returning new `AppService` with empty service list
+    /// marked as non-root.
     pub(crate) fn clone_config(&self) -> Self {
         AppService {
             config: self.config.clone(),
@@ -72,13 +74,13 @@ impl AppService {
         }
     }
 
-    /// Service configuration
+    /// Returns reference to configuration.
     #[must_use]
     pub fn config(&self) -> &AppConfig {
         &self.config
     }
 
-    /// Default resource
+    /// Returns default handler factory.
     #[must_use]
     pub fn default_service(&self) -> Rc<HttpNewService> {
         self.default.clone()
@@ -119,7 +121,7 @@ impl AppConfig {
         AppConfig { secure, host, addr }
     }
 
-    /// Needed in actix-test crate.
+    /// Needed in actix-test crate. Semver exempt.
     #[doc(hidden)]
     pub fn __priv_test_new(secure: bool, host: String, addr: SocketAddr) -> Self {
         AppConfig::new(secure, host, addr)
@@ -198,6 +200,7 @@ impl ServiceConfig {
     /// Add shared app data item.
     ///
     /// Counterpart to [`App::data()`](crate::App::data).
+    #[deprecated(since = "4.0.0", note = "Use `.app_data(Data::new(val))` instead.")]
     pub fn data<U: 'static>(&mut self, data: U) -> &mut Self {
         self.app_data(Data::new(data));
         self
@@ -263,6 +266,8 @@ mod tests {
     use crate::test::{call_service, init_service, read_body, TestRequest};
     use crate::{web, App, HttpRequest, HttpResponse};
 
+    // allow deprecated `ServiceConfig::data`
+    #[allow(deprecated)]
     #[actix_rt::test]
     async fn test_data() {
         let cfg = |cfg: &mut ServiceConfig| {