+++ src/data.rs
@@ -36,6 +36,11 @@ pub(crate) type FnDataFactory =
 /// If route data is not set for a handler, using `Data<T>` extractor would cause *Internal
 /// Server Error* response.
 ///
+// TODO: document `dyn T` functionality through converting an Arc
+// TODO: note equivalence of req.app_data<Data<T>> and Data<T> extractor
+// TODO: note that data must be inserted using Data<T> in order to extract it
+///
+/// # Examples
 /// ```
 /// use std::sync::Mutex;
 /// use actix_web::{web, App, HttpResponse, Responder};