+++ src/response/response.rs
@@ -52,7 +52,10 @@ impl HttpResponse<AnyBody> {
     #[inline]
     #[must_use]
     pub fn from_error(error: impl Into<Error>) -> Self {
-        error.into().as_response_error().error_response()
+        let error = error.into();
+        let mut response = error.as_response_error().error_response();
+        response.error = Some(error);
+        response
     }
 }