+++ CHANGES.md
@@ -2,8 +2,10 @@
 
 ## Unreleased - 2021-xx-xx
 ### Added
+* Add `ServiceRequest::parts_mut`. [#2177]
 * Add extractors for `Uri` and `Method`. [#2263]
-* Add extractor for `ConnectionInfo` and `PeerAddr`. [#2263]
+* Add extractors for `ConnectionInfo` and `PeerAddr`. [#2263]
+* Add `Route::service` for using hand-written services as handlers. [#2262]
 
 ### Changed
 * Change compression algorithm features flags. [#2250]
@@ -11,8 +13,10 @@
 * Add `#[must_use]` attribute for many public functions/methods. [#2281]
 
 
+[#2177]: https://github.com/actix/actix-web/pull/2177
 [#2250]: https://github.com/actix/actix-web/pull/2250
 [#2271]: https://github.com/actix/actix-web/pull/2271
+[#2262]: https://github.com/actix/actix-web/pull/2262
 [#2263]: https://github.com/actix/actix-web/pull/2263
 [#2281]: https://github.com/actix/actix-web/pull/2281